+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - [TD-1571] Elixir's Logger config will check for EX_LOGGER_FORMAT variable to override format
 
+### Changed
+
+- [TD-1530] Changed csv upload to write extra fields on metadata
+
 ## [2.15.0] 2019-03-18
 
 ### Changed